+++ src/nvim/os/input.c
@@ -19,6 +19,7 @@
 #include "nvim/memory.h"
 #include "nvim/msgpack_rpc/channel.h"
 #include "nvim/os/input.h"
+#include "nvim/profile.h"
 #include "nvim/screen.h"
 #include "nvim/state.h"
 #include "nvim/ui.h"
